Added appLevelTests and date is kept as object till printing.
What changes you have made?
Removed toJSON in getDate function and changed tests according to the
expected date.

Why this change is necessary?
Before this commit, the date is passed as string. So, if any requirement on
date further is given, it is hard.

What is the impact of the change?
Now, we can perform any operations on date.